initial page layout and file structure
