--- apps/web/.env.example
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_SERVER_URL=
