--- apps/web/src/components/providers.tsx
@@ -1,23 +1,0 @@
-"use client";
-
-import { ThemeProvider } from "./theme-provider";
-import { Toaster } from "./ui/sonner";
-
-
-export default function Providers({
-  children
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <ThemeProvider
-      attribute="class"
-      defaultTheme="system"
-      enableSystem
-      disableTransitionOnChange
-    >
-      {children}
-      <Toaster richColors />
-    </ThemeProvider>
-  );
-}
