--- bts.jsonc
@@ -1,25 +1,0 @@
-// Better-T-Stack configuration file
-// safe to delete
-
-{
-  "$schema": "https://r2.better-t-stack.dev/schema.json",
-  "version": "2.33.8",
-  "createdAt": "2025-08-15T15:32:19.830Z",
-  "database": "none",
-  "orm": "none",
-  "backend": "none",
-  "runtime": "none",
-  "frontend": [
-    "next"
-  ],
-  "addons": [
-    "turborepo",
-    "biome"
-  ],
-  "examples": [],
-  "auth": false,
-  "packageManager": "bun",
-  "dbSetup": "none",
-  "api": "none",
-  "webDeploy": "none"
-}
