+++ next.config.ts
@@ -1,0 +1,7 @@
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+	// Next.js 15+ uses App Router by default
+};
+
+export default nextConfig;
