+++ src/components/header.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { ModeToggle } from "./mode-toggle";
+
+export default function Header() {
+	return (
+		<div>
+			<div className="flex flex-row items-center justify-between px-2 py-1">
+				<ModeToggle />
+			</div>
+			<hr />
+		</div>
+	);
+}
