+++ src/components/providers.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import { ThemeProvider } from "./theme-provider";
+import { Toaster } from "./ui/sonner";
+
+export default function Providers({ children }: { children: React.ReactNode }) {
+	return (
+		<ThemeProvider
+			attribute="class"
+			defaultTheme="light"
+			enableSystem
+			disableTransitionOnChange
+		>
+			{children}
+			<Toaster richColors />
+		</ThemeProvider>
+	);
+}
